+++ recipes/redis_session.rb
@@ -1,7 +1,7 @@
 # Encoding: utf-8
 #
 # Cookbook Name:: magentostack
-# Recipe:: redis_base
+# Recipe:: redis_slave
 #
 # Copyright 2014, Rackspace US, Inc.
 #
@@ -18,6 +18,9 @@
 # limitations under the License.
 #
 
-stackname = 'magentostack'
-include_recipe 'stack_commons::redis_base'
-tag("#{stackname}-redis")
+bind_port = node['magentostack']['redis']['bind_port_session']
+server_name = "#{bind_port}-session-master"
+node.set['magentostack']['redis']['servers'][server_name] = { 'name' => server_name, 'port' => bind_port }
+MagentostackUtil.recompute_redis(node)
+tag('magentostack_redis')
+tag('magentostack_redis_session')