+++ recipes/redis_single.rb
@@ -17,4 +17,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-include_recipe 'stack_commons::redis_single' # ~RACK002
+# ensure redis is installed
+include_recipe 'redisio::install'
+
+bind_port = node['magentostack']['redis']['bind_port_single']
+server_name ="#{bind_port}-single-master"
+node.set['magentostack']['redis']['servers'][server_name] = { 'name' => server_name, 'port' => bind_port }
+MagentostackUtil.recompute_redis(node)
+tag('magentostack_redis')
+tag('magentostack_redis_single')